requirements.txt
python-dotenv
flask
azure-cognitiveservices-speech
